// src/lib/types/generated/models.ts
export interface CreateDraftRequest {
  artistKey: string;
  title: string;
  genres: string[];
  tracks: TrackRequest[];
  coverArt?: CoverArtRequest;
}
export interface TrackRequest {
  title: string;
  fileName: string;
  fileType: string;
  fileSize: number /* int64 */;
  duration: string;
  order: number /* int */;
}
export interface CoverArtRequest {
  fileName: string;
  fileType: string;
  fileSize: number /* int64 */;
}
export interface CreateDraftResponse {
  releaseHash: string;
  releaseKey: string;
  artistKey: string;
  presignedUrls: PresignedUrlsDTO;
}
export interface PresignedUrlsDTO {
  coverArt?: string;
  tracks: TrackUrlDTO[];
}
export interface TrackUrlDTO {
  hash: string;
  fileName: string;
  presignedUrl: string;
  storagePath: string;
}